requirements.txt
scikit-learn==0.24.2
pandas==1.1.5
mlflow==1.21.0
luigi==3.0.3
toml==0.10.2
python-dotenv==0.19.2